Use LDAP/AD stored avatar thumbnail in preference to Gravatar
Allow a user’s LDAP/AD stored avatar thumbnail to be queried and if found, used in preference to the Gravatar service.

Note for Enterprise Edition:

For EE, it'll work fine but only on the first defined LDAP provider. As noted in the code, I’m not sure the best way to resolve the users LDAP provider from within gravatar_service.rb but if someone can point me in the right direction then it'll be easy to support multiple LDAP providers. The alternative was to loop through all defined providers for a match but this seemed wasteful if the possibility existed to just resolve the users provider directly.