Fix underlying issue with emoji reactions
Issue was: blank space was rendering
as a element in an array of authors.
Element was being used by `join`.
Original fix !2450 was trying to remove the space after it happened.
This checks properly for it and only moves forward if it does not exist.
Also removes "me" upon unchecking emoji.