Merge branch 'fix_changelog' into 'master'
Move entry `Fix: As an admin, cannot add oneself as a member to a
group/project`

The CHANGELOG entry said it was fixed in 8.2.3. However, I found
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/1942 that
suggests it wasn't fixed until 8.3.0. A customer experienced this
issue on 8.2.3, still. This ultimately led to me investigating this
further.

[ci skip]

See merge request !2488