Merge branch 'rs-relax-autosize' into 'master'
Allow manual resize of js-autosize textareas

First, the autosize library was being too controlling and removed the
`resize` property from any elements to which it was attached, removing
the drag handle. We've disabled this behavior in the vendored library
and added a spec to prevent a regression during an upgrade.

Second, we detect (as best we can) when the user manually resizes an
autosize textarea, and then remove the autosize behavior from it and
increase its max-height.

This should allow for the best of both worlds.

Closes #12832

See merge request !2653