Merge branch 'backup-database-timeout-fix' into 'master'
Fix timeout issue for rake task gitlab:backup:create

This is a fix for a database timeout which can occur when the backup
create task is taking very long (1-2 hours).
It seems that ActiveRecord is loosing the connection after a hour
idleness and need to be reconnected before use.

See merge request !2757