Merge branch 'bugfix/jschatz1/sidebar-fixes/02-18-16' into 'master'
Uses cross browser niceScroll to do scrolling on sidebar.

Fixes: #13544 

Firefox does not support `-webkit-scrollbar` and there is no alternative for FF. So for example Gmail falls back to no fancy scrollbar. Stuck with jQuery, which is a good solution, since we already use it. 

This also leaves the default scrollbar as is. 

![Screen_Shot_2016-02-18_at_11.02.26_AM](/uploads/1184f06b71a8d9097ef1e3450a0c0ab3/Screen_Shot_2016-02-18_at_11.02.26_AM.png)

Also fixes horz scroll issue: 

![Screen_Shot_2016-02-18_at_11.09.56_AM](/uploads/63e9a8e824cac69a2745dd4dd84d1dfb/Screen_Shot_2016-02-18_at_11.09.56_AM.png)

cc @creamzy @dzaporozhets 

See merge request !2873