Merge branch 'fix/project-forks-page' into 'master'
workaround for forks with an invalid repo

Fixes https://gitlab.com/gitlab-org/gitlab-ce/issues/13465

The error occurs when a project returns a `nil` commit for a variety of reasons, I assume the repository is corrupt - perhaps as a result of a failed fork. 

With this MR, we do not show forks with corrupted repos in the list of forks, as this does not seem to work anyway. A better solution is to fix the cause of the issue and/or any data integrity problem...

Also fixing Sentry issue: 1180
`undefined method 'already_forked?' for nil:NilClass `


See merge request !2836