Remove redundant integration tests
These three tests were essentially checking that adding a note to
something updated its `noteable`'s `updated_at` attribute.

This is well-tested Rails behavior and we shouldn't feel the need to
write an integration test to verify it. At most we should be ensuring
that the association definition adds the `touch: true` option, which we
now do in Note's unit test.