Merge branch 'dashboard-labels' into 'master'
Dashboard labels

Previously because there were no JSON endpoint for labels or milestones
it was fetching HTML and parsing that. This is wrong.

It now fetches from a JSON endpoint.

This also fixes the dashboard/issues page not loading labels or
milestones as the path used to be only for a single project. So instead
I have created a endpoint for labels & milestones JSON on dashboard.

See merge request !3253