Only rename the column if its not there yet
The problem can only arise when upgrading from a CE version to an
 EE version after last months release as in that release the table
appearances was created. But when it was created we already took
care of this rename. Thus on an upgrade the migration will fail as
the column `light_logo` is not in the table, and `header_logo` already
is.

Thus this check solve this.