Upgrade Ruby 2.2 tests to 2.3
This changes the GitLab CI Tests that test compatibility with Ruby 2.2
to instead test compat with Ruby 2.3.

We’ve discussed skipping straight to 2.3.0 when next upgrading the
required Ruby version.

Since Rails 5 requires 2.2.2, and we have to upgrade anyway, may as
well do it now and not risk having to upgrade again come Rails 5.1 or
5.2.

Test failures on Ruby 2.3 do not fail the build.