In emails-on-push message, indicate message that individual file diff is too large
Diff overflow limits apply to different cases:

1. The entire diff exceeds 100 files OR 5000 lines
2. One single file exceeds 100 KB

Example case 2: Let's say we have a diff of 50 files, most of them under 100
KB, but one of them over that limit. In this case, the idea of
gitlab-org/gitlab-ce!2705 is that GitLab should show the diff of the 49 files
but omit the last one.