Add License Finder gem.
In order to rehost all our gems in our own gem host, we need to have the legal rights to do so for every gem should they be taken down from RubyGems. License Finder automates checking of gems to ensure that we're in the clear legally.

Approved the MIT License because it essentially allows us to do "whatever" with those gems.

I am not a lawyer.

https://github.com/pivotal/LicenseFinder