Updated list of InfluxDB queries/config
This setup is quite a bit different from before. In the previous setup
raw data was kept around for 30 days and downsampled data for 7 days.
This became problematic for GitLab.com as the number of points and
series resulted in InfluxDB running out of memory when starting up
(besides taking up 30 GB of storage).

To work around this the new setup keeps raw data around for _only_ an
hour while keeping downsampled data around for 7 days. In turn all
Grafana dashboards _only_ query the downsampled data instead of also
querying raw data.

Based on rough calculations this setup needs around 2GB of storage for 1
week of data, excluding whatever is needed for storing the raw data
(this highly depends on the amount of traffic).

If users want to use this new setup they have to remove any existing
dashboards provided by GitLab.com and re-import the ones from the
Grafana dashboards repository
(https://gitlab.com/gitlab-org/grafana-dashboards/). Should users wish
to change their default retention policy the easiest way of doing so is
to simply drop the database and re-run the InfluxDB commands added by
this commit. Users who want to keep their default retention policy as-is
can simply create the "downsampled" policy and run the other commands.