Remove unused .contributed-projects class
While working on #13401 and trying to add a new tab to the user profile
page, I came across this. I noticed that the `contributed-projects`
class was only being used in order to select the div in the tests. For
consistency with the other tabs, I decided to remove this class and use
the div's id for the selector.