Merge branch 'fix-team-build-state-in-mr-widget' into 'master'
Merge request widget displays TeamCity build state and code coverage correctly again

## What does this MR do?
This MR contains a fix for a regression introduced in `8.7`. In former version, the TeamCity build status was always displayed correctly. In `8.7` the build state is still checked, but the UI is not updated correctly any longer.

## Are there points in the code the reviewer needs to double check?
The changes are quite simple, so please simply double check them.

## Why was this MR needed?
This MR is needed to make the TeamCity build status working again.

## What are the relevant issue numbers?
#17080 

See merge request !3998