Re-added user checks in Participable#participants
The old code would check if every returned user could read the current
project. In the new setup this is already done for the mentioned users
and using the old approach would lead to the code running lots of
useless SQL queries. To work around this (while still ensuring every
user has the right permissions) we _only_ do an extra check for users
returned directly by attributes, skipping users returned by a Proc.

When grabbing the participants of an issue with 200 mentioned users this
increases the query count from around 12 to 30 (around 48 when the
issue/merge request has both an author and assignee). If we were to
simply check permissions for every user the query count would be between
1000 and 1100.